# api/internal/swagger/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: API for managing Allure test reports.
  title: AllureDeck API
  version: 2.0.0
paths:
...
  /projects/{project_id}/jobs/{job_id}:
    get:
      description: Returns the current status of an async report generation job.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get async report generation job status
      tags:
      - reports
...
  /projects/{project_id}/reports:
    get:
      description: Returns paginated metadata for all generated reports of a project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get report history
      tags:
      - reports
    post:
      description: Triggers Allure report generation for the specified project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Execution name (CI provider, e.g. 'GitHub Actions')
        in: query
        name: execution_name
        type: string
      - description: Execution from (CI build URL)
        in: query
        name: execution_from
        type: string
      - description: Execution type
        in: query
        name: execution_type
        type: string
      - description: Store results (true/1)
        in: query
        name: store_results
        type: string
      - description: CI branch name
        in: query
        name: ci_branch
        type: string
      - description: CI commit SHA
        in: query
        name: ci_commit_sha
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Generate a report
      tags:
      - reports
  /projects/{project_id}/reports/{report_id}:
    delete:
      description: Removes a single numbered report directory for a project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a report
      tags:
      - reports
...
  /projects/{project_id}/reports/{report_id}/categories:
    get:
      description: Returns the failure categories from the categories widget JSON
        for a given report.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID or 'latest'
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get report defect categories
      tags:
      - reports
  /projects/{project_id}/reports/{report_id}/environment:
    get:
      description: Returns the environment properties from the environment widget
        JSON for a given report.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Report ID or 'latest'
        in: path
        name: report_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get report environment info
      tags:
      - reports
...
  /projects/{project_id}/reports/history:
    delete:
      description: Removes all historical report snapshots for a project.
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Clean report history
      tags:
      - reports
  /projects/{project_id}/reports/history/group:
    delete:
      description: Deletes all report history for all child projects of the given
        parent.
      parameters:
      - description: Parent project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Clean history for all projects in a group
      tags:
      - reports
...
swagger: "2.0"